+++ client/src/components/SearchBarArea/index.jsx
@@ -1,5 +1,5 @@
 import React, {useState} from "react";
-
+import { Link } from 'react-router-dom'
 const BarSearch =() => {
     const [searchArea, setSearchArea] = useState('');
 
@@ -21,8 +21,10 @@ const BarSearch =() => {
                 onChange={handleInputChange}
                 placeholder="Search Location"
                 />
+                <Link to={'/area-info'}>
                 <button type="submit">Search</button>
-        </form>
+                </Link>
+            </form>
     );
 };
