+++ client/src/components/pages/AreaInfo/AreaInfo.jsx
