+++ client/src/components/pages/PlantInfo/PlantInfo.jsx
